llama-cpp-server/requirements.txt
numpy
fastapi[standard]
uvicorn
ninja
llama-cpp-python

